add pdfkit

+++ api/src/models/billModel.js
@@ -3,8 +3,8 @@ import sequelize from '../../db';
 
 const Bill = sequelize.define('bill', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  incomeTax: { type: DataTypes.DECIMAL, allowNull: false },
-  militaryTax: { type: DataTypes.DECIMAL, allowNull: false },
+  incomeTax: { type: DataTypes.DOUBLE, allowNull: false },
+  militaryTax: { type: DataTypes.DOUBLE, allowNull: false },
 });
 
 export default Bill;
